+++ core-tests/shared/src/test/scala/zio/FiberRuntimeSpec.scala
@@ -1,6 +1,5 @@
 package zio
 
-import zio.Clock.ClockLive
 import zio.internal.FiberScope
 import zio.test._
 
@@ -63,30 +62,8 @@ object FiberRuntimeSpec extends ZIOBaseSpec {
             }
         }
       }
-    ),
-    suite("supervisor")(
-      suite("onStart and onEnd are called exactly once")(
-        test("sync effect") {
-          for {
-            s <- ZIO.succeed(new StartEndTrackingSupervisor)
-            f <- ZIO.unit.fork.supervised(s)
-            _ <- f.await
-            // onEnd might be called after the forked fiber notifies the current fiber
-            _ <- ZIO.succeed(s.onEndCalls).repeatUntil(_ > 0)
-          } yield assertTrue(s.onStartCalls == 1, s.onEndCalls == 1)
-        },
-        test("async effect") {
-          for {
-            s <- ZIO.succeed(new StartEndTrackingSupervisor)
-            f <- ClockLive.sleep(100.micros).fork.supervised(s)
-            _ <- f.await
-            // onEnd might be called after the forked fiber notifies the current fiber
-            _ <- ZIO.succeed(s.onEndCalls).repeatUntil(_ > 0)
-          } yield assertTrue(s.onStartCalls == 1, s.onEndCalls == 1)
-        }
-      ) @@ TestAspect.nonFlaky(100)
     )
-  ) @@ TestAspect.timeout(10.seconds)
+  )
 
   private final class YieldTrackingSupervisor(
     latch: Promise[Nothing, Unit],
@@ -114,30 +91,6 @@ object FiberRuntimeSpec extends ZIOBaseSpec {
       if (onEndCalled) latch.unsafe.done(Exit.unit) // onEnd gets called before onSuspend
       ()
     }
-  }
-
-  private final class StartEndTrackingSupervisor extends Supervisor[Unit] {
-    private val _onStartCalls, _onEndCalls = new AtomicInteger(0)
-
-    def value(implicit trace: Trace): UIO[Unit] = ZIO.unit
-
-    def onStart[R, E, A](
-      environment: ZEnvironment[R],
-      effect: ZIO[R, E, A],
-      parent: Option[Fiber.Runtime[Any, Any]],
-      fiber: Fiber.Runtime[E, A]
-    )(implicit unsafe: Unsafe): Unit = {
-      _onStartCalls.incrementAndGet()
-      ()
-    }
-
-    def onEnd[R, E, A](value: Exit[E, A], fiber: Fiber.Runtime[E, A])(implicit unsafe: Unsafe): Unit = {
-      _onEndCalls.incrementAndGet
-      ()
-    }
-
-    def onStartCalls = _onStartCalls.get
-    def onEndCalls   = _onEndCalls.get
   }
 
 }
