+++ core-tests/shared/src/test/scala/zio/SupervisorSpec.scala
@@ -1,7 +1,10 @@
 package zio
 
+import zio.Clock.ClockLive
 import zio.test.TestAspect.{exceptJS, nonFlaky}
 import zio.test._
+
+import java.util.concurrent.atomic.AtomicInteger
 
 object SupervisorSpec extends ZIOBaseSpec {
 
@@ -31,7 +34,27 @@ object SupervisorSpec extends ZIOBaseSpec {
       DifferSpec.diffLaws(Differ.supervisor)(genSupervisor)((left, right) =>
         Supervisor.toSet(left) == Supervisor.toSet(right)
       )
-    }
+    },
+    suite("onStart and onEnd are called exactly once")(
+      test("sync effect") {
+        for {
+          s <- ZIO.succeed(new StartEndTrackingSupervisor)
+          f <- ZIO.unit.fork.supervised(s)
+          _ <- f.await
+          // onEnd might be called after the forked fiber notifies the current fiber
+          _ <- ZIO.succeed(s.onEndCalls).repeatUntil(_ > 0)
+        } yield assertTrue(s.onStartCalls == 1, s.onEndCalls == 1)
+      },
+      test("async effect") {
+        for {
+          s <- ZIO.succeed(new StartEndTrackingSupervisor)
+          f <- ClockLive.sleep(100.micros).fork.supervised(s)
+          _ <- f.await
+          // onEnd might be called after the forked fiber notifies the current fiber
+          _ <- ZIO.succeed(s.onEndCalls).repeatUntil(_ > 0)
+        } yield assertTrue(s.onStartCalls == 1, s.onEndCalls == 1)
+      }
+    ) @@ TestAspect.nonFlaky(100)
   )
 
   val genSupervisor: Gen[Any, Supervisor[Any]] =
@@ -105,4 +128,29 @@ object SupervisorSpec extends ZIOBaseSpec {
           promise.unsafe.done(ZIO.succeed(ref.unsafe.get))
       }
     }
+
+  private final class StartEndTrackingSupervisor extends Supervisor[Unit] {
+    private val _onStartCalls, _onEndCalls = new AtomicInteger(0)
+
+    def value(implicit trace: Trace): UIO[Unit] = ZIO.unit
+
+    def onStart[R, E, A](
+      environment: ZEnvironment[R],
+      effect: ZIO[R, E, A],
+      parent: Option[Fiber.Runtime[Any, Any]],
+      fiber: Fiber.Runtime[E, A]
+    )(implicit unsafe: Unsafe): Unit = {
+      _onStartCalls.incrementAndGet()
+      ()
+    }
+
+    def onEnd[R, E, A](value: Exit[E, A], fiber: Fiber.Runtime[E, A])(implicit unsafe: Unsafe): Unit = {
+      _onEndCalls.incrementAndGet
+      ()
+    }
+
+    def onStartCalls = _onStartCalls.get
+    def onEndCalls   = _onEndCalls.get
+  }
+
 }
